API Reference

DqdLeadsCavity.D_sop — Method

Lindblad dissipator superoperator

DqdLeadsCavity._integrand_heat_current_neqgf — Method

Particle current integrand for NEGFs approach

DqdLeadsCavity._integrand_particle_current_avg_neqgf — Method

Particle current integrand for NEGFs approach

DqdLeadsCavity._is_real — Method

Helper function to check whether a number is approximately real

DqdLeadsCavity._transmission_dqd — Method

Transmission function of the DQD under the wideband approximation [eq. (B5) Potts2021] ≡ [eq. (A49) Prech2023]

Arguments

ω::Real: Energy (integration variable)

DqdLeadsCavity.build_H_JC — Method

Hamiltonian for the DQD-cavity interaction in the JC model

DqdLeadsCavity.build_H_cav_drive_coh — Method

Hamiltonian for the coherent cavity drive

DqdLeadsCavity.build_H_dqd_LR — Method

DQD Hamiltonian in the L-R basis

DqdLeadsCavity.build_H_dqd_ge — Method

DQD Hamiltonian in the g-e parametrization

DqdLeadsCavity.build_L_ops_dqd_gl — Method

Lindblad dissipator for the global approach according to [eq. (88) Potts2021]. This is valid for the non-interacting DQD case (U = 0).

DqdLeadsCavity.build_L_ops_dqd_loc — Method

Lindblad dissipator for the non-interacting (U = 0) DQD according to the local approach [eq. (92) Potts2021].

DqdLeadsCavity.build_L_ops_semilocal_LR — Method

Lindblad operators corresponding to the semi-local LME evaluating the fermi distributions at the corresponding on-site energies.

DqdLeadsCavity.build_L_ops_thcg_int — Method

Lindblad operators for the interacting DQD for the THC global case according to eq. (118) [Potts2021]. This only works for ϵL = ϵR.

DqdLeadsCavity.build_cav_a_op — Method

Cavity annihilation operator

DqdLeadsCavity.build_dqd_basis_LR — Method

DQD L-R ket basis. We implicitly assume the ordering convention dR'dL'|0>=|1,1>.

DqdLeadsCavity.build_dqd_basis_ge — Method

DQD ket basis in the ground-excited basis.

DqdLeadsCavity.build_dqd_fermi_ops_LR — Method

DQD vacuum ladder operators in the left-right basis

DqdLeadsCavity.build_dqd_fermi_ops_ge — Method

DQD fermionic operators in the ground-excited basis

DqdLeadsCavity.build_dqd_ladder_ops_ge — Method

DQD ground-excited ladder operators

DqdLeadsCavity.build_dqd_number_op — Method

DQD total number operator

DqdLeadsCavity.build_dqd_number_ops_LR — Method

DQD number operators in the LR basis

DqdLeadsCavity.build_dqd_number_ops_ge — Method

DQD number operators in the g-e basis

DqdLeadsCavity.build_dqd_σz_ge — Method

DQD-σz operator in the ground-excited basis

DqdLeadsCavity.build_dqd_σz_op — Method

DQD σz-operator in the ground-excited basis

DqdLeadsCavity.build_id_cavity — Method

Cavity Identity operator

DqdLeadsCavity.build_id_dqd — Method

DQD Identity operator

DqdLeadsCavity.build_σz_ge — Method

σz in the g-e parametrization

DqdLeadsCavity.context_ge — Method

Helper function to gather g-e labelled quantities

DqdLeadsCavity.current_output — Method

Calculate the average of the output stochastic current according to eq. (40) of [Landi et. al. 2024]

Arguments

ρ::QuantumObject{Operator}: Reduced density matrix of the system

L_ops::Vector{QuantumObject{Operator}}}: Vector of Lindblad operators

Keyword Arguments

ν_vec::Vector{Int64}: Weights corresponding to each Lindblad operator in L_ops

Returns

I_avg::Float64: Output average stochastic current

DqdLeadsCavity.fermi — Method

Fermi distribution

Arguments

ϵ::Real: Energy

μ::Real: Chemical potential

T::Real: Temperature

DqdLeadsCavity.get_chemical_potentials — Method

Calculates chemical potentials of the leads

DqdLeadsCavity.get_coherence_LR_gl — Method

Analytical steady-state solution for the coherence between the ground and excited levels of the DQD according to the global approach [eq. (A29) Prech2023]

DqdLeadsCavity.get_coherence_LR_num — Method

Numerical coherence between left-right states of the DQD

DqdLeadsCavity.get_coherence_gl_ana — Method

Analytical coherence for the global approach according to [eq. (A.29) Prech2023]

DqdLeadsCavity.get_concurrence_LR_num — Method

Numerical concurrence in the

DqdLeadsCavity.get_concurrence_gl_ana — Method

Analytical concurrence for the global approach according to [eq. (A.30) Prech2023]

DqdLeadsCavity.get_coupling_strengths_gl — Method

Coupling strengths for the global approach according to [eq. (A24) Prech2023] ≡ [eq. (89) Potts2021]

DqdLeadsCavity.get_current_heat_neqgf — Method

Compute steady-state heat current through the non-interacting DQD using NEQGFs according to eq. (B.3) [Potts et. al. 2021])

Returns

I_avg::Real

DqdLeadsCavity.get_dim — Method

Get DQD Hilbert space dimension

DqdLeadsCavity.get_eigen_energies — Method

DQD eigen-energies

DqdLeadsCavity.get_fermi_factors_gl — Method

Fermi factors for the global approach [Potts2021]

DqdLeadsCavity.get_fermi_factors_loc — Method

Fermi factors for the local approach [Potts2021]

DqdLeadsCavity.get_heat_current_gl — Method

Analytical steady-state solution for the heat current according to the global approach [eq. (A33) Prech2023] or [eq. (B.8) Potts2021]

DqdLeadsCavity.get_heat_current_loc — Method

Aalytical steady-state heat current through the non-interacting DQD according to the local approach [eq. (B.6) Potts et. al. 2021]

DqdLeadsCavity.get_heat_current_num — Method

Numerical heat current calculation

DqdLeadsCavity.get_n_cav_ss — Method

Steady-state photon number in the cavity without DQD

DqdLeadsCavity.get_nth_cav — Method

Thermal occupation at the cavity frequency

DqdLeadsCavity.get_occupation_LR_gl — Method

Analytical steady-state solution for the occupation of the DQD left and right states according to global approach [eq. (A28) Prech2023]

DqdLeadsCavity.get_occupation_ge_gl_ana — Method

Analytical steady-state solution for the occupation of the DQD grond/excited state according to the global approach[eq. (A27) Prech2023]

DqdLeadsCavity.get_onsite_energies — Method

DQD on-site energies

DqdLeadsCavity.get_particle_current_gl — Method

Analytical steady-state particle current for the non-interacting DQD according to the global approach [eq. (A31) Prech2023]. Due to particle conservation IR = -IL.

DqdLeadsCavity.get_particle_current_loc — Method

Analytical steady-state particle current for the non-interacting DQD according to the local approach[eq. (A12) Prech2023]. Due to conservation of particles IR = -IL

DqdLeadsCavity.get_particle_current_neqgf — Method

Compute steady-state particle current through the non-interacting DQD using NEQGFs according to [eq. (14) Prech et. al. 2023]

Returns

I_avg::Real

DqdLeadsCavity.get_particle_current_num — Method

Numerical particle current calculation

DqdLeadsCavity.get_populations_LR_num — Method

Numerical populations of the DQD in the left-right basis

DqdLeadsCavity.get_Δc — Method

DQD-cavity detuning

DqdLeadsCavity.get_Δd — Method

DQD-driving detuning

DqdLeadsCavity.get_Ω — Method

DQD energy split in the g-e basis

DqdLeadsCavity.get_α — Method

Cavity displacement field

DqdLeadsCavity.get_θ — Method

DQD mixing angle in the g-e basis. Sign convention follows [Prech2023] (ϵL < ϵR).

DqdLeadsCavity.id_no_vacuum — Method

Identity for the subspace without the vacuum state