API Reference

DqdLeadsCavity.build_H_dqd_LR — Method

Build DQD Hamiltonian in the L-R basis

DqdLeadsCavity.build_H_dqd_ge — Method

Build DQD Hamiltonian in the g-e parametrization

DqdLeadsCavity.build_L_ops_local_LR — Method

Build Lindblad operators corresponding to the local LME evaluating the fermi distributions at the corresponding on-site energies.

DqdLeadsCavity.build_L_ops_local_ge — Method

Build Lindblad operators corresponding to the local LME evaluating the eigen-energies of the DQD (g-e)

DqdLeadsCavity.build_L_ops_semilocal_LR — Method

Build Lindblad operators corresponding to the semi-local LME evaluating the fermi distributions at the corresponding on-site energies.

DqdLeadsCavity.build_dqd_basis_LR — Method

Build DQD L-R ket basis

DqdLeadsCavity.build_dqd_basis_ge — Method

Build DQD g-e ket basis

DqdLeadsCavity.build_dqd_ladder_ops_LR — Method

Build DQD L-R ladder operators

DqdLeadsCavity.build_dqd_ladder_ops_ge — Method

Build DQD g-e ladder operators

DqdLeadsCavity.build_dqd_number_ops_LR — Method

Build DQD number operators in the LR basis

DqdLeadsCavity.build_σz_ge — Method

Build σz in the g-e parametrization

DqdLeadsCavity.fermi — Method

Fermi distribution

Arguments

ϵ::Real: Energy

μ::Real: Chemical potential

T::Real: Temperature

DqdLeadsCavity.get_dim — Method

Get object dimension

DqdLeadsCavity.get_eigen_energies — Method

DQD eigen-energies

DqdLeadsCavity.get_onsite_energies — Method

DQD on-site energies

DqdLeadsCavity.get_Δd — Method

Detuning between the DQD and cavity drive

DqdLeadsCavity.get_Ω — Method

DQD energy split in the g-e basis

DqdLeadsCavity.get_θ — Method

DQD mixing angle in the g-e basis

DqdLeadsCavity.id_no_vacuum — Method

Identity for the subspace without the vacuum state